Refactor card loading logic to be a more efficient. Specifically: invoke CardAPI once per panel load.
