+++ app/src/components/GamePanel.js
@@ -9,22 +9,34 @@ import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 import {withStyles} from '@material-ui/styles';
 import Typography from '@material-ui/core/Typography';
+import CardAPI from '../api/CardAPI';
 
 class GamePanel extends Component {
+  state = {
+    showName: false,
+    names: [],
+    cards: [],
+    score: 0,
+  }
+
   constructor(props) {
     super(props);
-    let namesArray = [];
-    for (let i = 0; i < props.cards.length; i++) {
-      namesArray.push({name: undefined});
-    }
-    this.state = {
-      showName: false,
-      names: namesArray,
-      score: 0,
-    };
-
     this.onNameChange = this.onNameChange.bind(this);
     this.onNameKeyDown = this.onNameKeyDown.bind(this);
+  }
+
+  setCards(cards) {
+    this.setState({
+      cards: [...cards],
+      names: cards.map( (card) => Object.assign({name:undefined}) )
+    });
+  }
+
+  componentDidMount() {
+    return CardAPI.buildCards(this.props.config.data.memberlist)
+      .then(CardAPI.randomizeOrder)
+      .then(cards => this.setCards(cards))
+      .catch(error => console.error(error));
   }
 
   /**
@@ -45,7 +57,7 @@ class GamePanel extends Component {
    */
   onNameKeyDown(index, e) {
     if (e.key === 'Enter') {
-      let answer = this.props.cards[index].name.toUpperCase().split(' ')[0];
+      let answer = this.state.cards[index].name.toUpperCase().split(' ')[0];
       let guess = e.target.value.toUpperCase().split(' ')[0];
       let guessResult = answer === guess;
 
@@ -88,7 +100,7 @@ class GamePanel extends Component {
 
         {/* Valid cards are show here */}
         <Grid container spacing={4}>
-          {this.props.cards
+          {this.state.cards
             .filter(card => !card.error)
             .map((card, index) => (
               <Grid item key={card.name} xs={12} sm={6} md={4}>
@@ -139,7 +151,7 @@ class GamePanel extends Component {
 
         {/* Cards with errors are shown here */}
         <Grid container spacing={4}>
-          {this.props.cards
+          {this.state.cards
             .filter(card => !!card.error)
             .map(card => (
               <Grid item key={card.name} xs={12} sm={6} md={4}>
